+++ js/api/posts/new-posts/create-post.mjs
@@ -1,4 +1,4 @@
-import { API_BASE_URL } from "../base-url.mjs";
+import { API_BASE_URL } from "../../base-url.mjs";
 
 export async function createPost(title, body, media) {
   try {
