+++ js/feed.mjs
@@ -1,7 +1,7 @@
 import { fetchPosts } from "./api/posts/fetch-posts.mjs";
 import { renderPosts } from "./api/posts/render-posts.mjs";
-import { addFilterEventListener } from "./api/posts/filter-posts.mjs";
-import { addSearchEventListener } from "./api/posts/search-posts.mjs";
+import { addFilterEventListener } from "./api/posts/filter-and-search/filter-posts.mjs";
+import { addSearchEventListener } from "./api/posts/filter-and-search/search-posts.mjs";
 
 // Define viewPost function
 export function viewPost(postId) {
